pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["gendiff"]

[project]
name = "hexlet-code"
version = "0.1.0"
description = "Compares two configuration files and shows a difference"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    'prompt>=0.4.1',
    'argparse',
    'pyyaml'
]

[project.scripts]
gendiff = "gendiff.cli:main"

[dependency-groups]
dev = [
    "ruff>=0.9.1",
]

[tool.ruff]
line-length = 80

[tool.ruff.lint]
preview = true
select = ["E", "F", "I", "C90"]

[tool.lint.per-file-ignores]
"__init__.py" = ["F401"]